requirements.txt
mcp>=0.11.0
azure-kusto-data>=4.2.0
azure-kusto-ingest>=4.2.0
azure-identity>=1.15.0
python-dotenv>=1.0.0